New optional addon: "JDScriptLauncher"
It will add "shortcuts" to the addon menu bar to execute external scripts (batch, bash or whatever - but must be executable!).
The scripts must exist in the folder "scripts" in the JD root. Note that this folder will be created if it does not exist yet.
This addon will not search for scripts recursively.