Mac Folder Installer changes.
The Folder Installer does not wrap all contents in the JDownloader2.app, but (like on windows and linux) in the same folder. 
We have to update some functions that work on the old *.app->Contents->.... structure.



refs #54757
Prepared the validate Method to return a special exception if the filepaths is too long.

build>full